refactor: 패키지명 data -> persistence 변경
