+++ src/test/java/roomescape/persistence/dao/JdbcPlayPlayTimeDaoTest.java
@@ -1,4 +1,4 @@
-package roomescape.data.dao;
+package roomescape.persistence.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -13,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.business.domain.PlayTime;
-import roomescape.data.entity.PlayTimeEntity;
+import roomescape.persistence.entity.PlayTimeEntity;
 
 @JdbcTest
 class JdbcPlayPlayTimeDaoTest {
