+++ src/test/java/roomescape/persistence/dao/JdbcReservationDaoTest.java
@@ -1,4 +1,4 @@
-package roomescape.data.dao;
+package roomescape.persistence.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -14,8 +14,8 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import roomescape.business.domain.PlayTime;
 import roomescape.business.domain.Reservation;
-import roomescape.data.entity.ReservationEntity;
-import roomescape.data.entity.PlayTimeEntity;
+import roomescape.persistence.entity.ReservationEntity;
+import roomescape.persistence.entity.PlayTimeEntity;
 
 @JdbcTest
 class JdbcReservationDaoTest {
